src/requirements.txt
pydantic
python-dotenv
requests
rich
